network posix: Connect using right protocol family
Add an hint when calling `getaddrinfo()` to show that we are after a connection
that supports the `SOCK_STREAM` type.

Check the return value of `getaddrinfo()`.

Create a socket that has a family that matches the address we try to setup a
connection to.

Use a for-loop to simplify the code a bit.